requirements.txt
addict==2.4.0
affine==2.4.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
altair==5.4.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
anywidget==0.9.13
asttokens==2.4.1
asyncio==3.4.3
attrs==24.2.0
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.3.3
azure-ai-ml==1.19.0
azure-ai-translation-text==1.0.1
azure-cognitiveservices-speech==1.40.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-maps-render==1.0.0b2
azure-maps-search==1.0.0b2
azure-mgmt-core==1.4.0
azure-mgmt-maps==2.1.0
azure-storage-blob==12.22.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
beautifulsoup4==4.12.3
blinker==1.8.2
bqplot==0.12.43
branca==0.7.2
cachelib==0.9.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.5
colorama==0.4.6
colour==0.1.5
comm==0.2.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
duckdb==1.0.0
executing==2.0.1
faiss-cpu==1.8.0.post1
filelock==3.15.4
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==4.0.1
flask-restx==1.3.0
folium==0.17.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
-e git+https://github.com/TensorTeen/ISRO-Hackathon-2024.git@842dd38619d6b12cee460760dcd0642bad5a3857#egg=geoawaregpt
geojson==3.1.0
geopandas==1.0.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
groundingdino-py==0.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipython==8.26.0
ipytree==0.2.2
ipyvue==1.11.1
ipyvuetify==1.10.0
ipywidgets==8.1.3
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.2.13
langchain-community==0.2.12
langchain-core==0.2.30
langchain-openai==0.1.21
langchain-text-splitters==0.2.2
langsmith==0.1.99
leafmap==0.36.6
litellm==1.43.9
localtileserver==0.10.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
morecantile==5.3.1
mpmath==1.3.0
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.4.0
networkx==3.3
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.40.6
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
patool==2.3.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
portalocker==2.10.1
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg2==2.9.9
psygnal==0.11.1
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydash==8.0.3
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.8.3
pytest==8.3.2
python-box==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
rasterio==1.3.10
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rio-cogeo==5.3.3
rio-tiler==6.6.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scooby==0.10.0
segment-anything-hq==0.3
segment-anything-py==1.0.1
segment-geospatial==0.10.7
server-thread==0.2.0
setuptools==72.2.0
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snuggs==1.4.7
soupsieve==2.6
SQLAlchemy==2.0.32
stack-data==0.6.3
streamlit==1.37.1
strictyaml==1.7.3
supervision==0.6.0
sympy==1.13.2
tenacity==8.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
transformers==4.44.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.3
whitebox==2.3.5
whiteboxgui==2.3.0
widgetsnbextension==4.0.11
xarray==2024.7.0
xyzservices==2024.6.0
yapf==0.40.2
yarl==1.9.4
zipp==3.20.0
